app/requirements.txt
aerich==0.6.2
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.4.1
asyncpg==0.25.0
boto3==1.20.37
botocore==1.23.37
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
ddlparse==1.10.0
dictdiffer==0.9.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.72.0
h11==0.12.0
httptools==0.2.0
idna==3.3
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
orjson==3.6.5
print-logger==1.0.3
pydantic==1.9.0
pyparsing==3.0.6
pypika-tortoise==0.1.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
python-rapidjson==1.5
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
s3transfer==0.5.0
sentry-sdk==1.5.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomlkit==0.8.0
tortoise-orm==0.18.1
typing-extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
